+++ battlefox_database/src/adkats.rs
@@ -1,0 +1,2 @@
+pub mod players;
+pub mod bans;
